+++ ext/monitor/src/main/java/app/monitor/job/Collector.java
@@ -1,0 +1,10 @@
+package app.monitor.job;
+
+import core.framework.internal.stat.Stats;
+
+/**
+ * @author neo
+ */
+public interface Collector {
+    Stats collect();
+}